requirements.txt
# NutriAI v4 — requirements
streamlit>=1.32.0
anthropic>=0.25.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai
Pillow
# sqlite3 is stdlib — no install needed
